refactor(number): 클래스 이름 변경
- 추상 클래스를 Number로 변경
- Lotto 객체에 포함되는 객체를 LottoNumber로 변경
