+++ src/main/java/lotto/domain/lottoForm/Lotto.java
@@ -1,7 +1,7 @@
 package lotto.domain.lottoForm;
 
-import lotto.domain.number.LottoNumber;
 import lotto.domain.number.Number;
+import lotto.domain.number.LottoNumber;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -9,25 +9,25 @@ import java.util.stream.Collectors;
 public class Lotto extends LottoForm {
     private final String DELIMITER = ", ";
     private final List<Integer> numbers;
-    private final List<Number> numbers2;
+    private final List<LottoNumber> numbers2;
 
     public Lotto(List<Integer> numbers) {
         this.numbers = validateAndSort(numbers);
         this.numbers2 = validateAndSort(numbers).stream()
-                .map(Number::new)
+                .map(LottoNumber::new)
                 .toList();
     }
 
     @Override
     public String toString() {
         String result = numbers2.stream()
-                .map(LottoNumber::getNumber)
+                .map(Number::getNumber)
                 .map(String::valueOf)
                 .collect(Collectors.joining(DELIMITER));
         return String.format("[%s]", result);
     }
 
-    public List<Number> getNumbers() {
+    public List<LottoNumber> getNumbers() {
         return numbers2;
     }
 
@@ -37,7 +37,7 @@ public class Lotto extends LottoForm {
                 .count());
     }
 
-    public boolean hasBonusNumber(LottoNumber bonusNumber) {
+    public boolean hasBonusNumber(Number bonusNumber) {
         return numbers2.stream()
                 .anyMatch(bonusNumber::isSame);
     }
