+++ src/main/java/lotto/domain/lottoForm/WinningNumbers.java
@@ -1,8 +1,8 @@
 package lotto.domain.lottoForm;
 
 import lotto.domain.number.BonusNumber;
-import lotto.domain.number.LottoNumber;
 import lotto.domain.number.Number;
+import lotto.domain.number.LottoNumber;
 
 import java.util.Arrays;
 import java.util.List;
@@ -12,12 +12,12 @@ import static lotto.message.ErrorMessage.NON_INTEGER_LOTTO;
 
 public class WinningNumbers extends LottoForm {
     private final String SPLIT_DELIMITER = ",";
-    private final List<Number> numbers2;
+    private final List<LottoNumber> numbers2;
 
     public WinningNumbers(String input) {
         List<Integer> rawNumbers = convertToIntegers(input);
         this.numbers2 = validateAndSort(rawNumbers).stream()
-                .map(Number::new)
+                .map(LottoNumber::new)
                 .toList();
     }
 
@@ -34,15 +34,15 @@ public class WinningNumbers extends LottoForm {
 
     public void validateDuplicate(BonusNumber bonusNumber) {
         numbers2.stream()
-                .filter(number2 -> number2.isSame(bonusNumber))
+                .filter(lottoNumber2 -> lottoNumber2.isSame(bonusNumber))
                 .findAny()
-                .ifPresent(foundNumber -> {
+                .ifPresent(foundLottoNumber -> {
                     throw new IllegalArgumentException(BONUS_NUMBER_DUPLICATE.getMessage());
                 });
     }
 
-    public boolean contains(LottoNumber number) {
+    public boolean contains(Number number) {
         return numbers2.stream()
-                .anyMatch(number2 -> number2.isSame(number));
+                .anyMatch(lottoNumber2 -> lottoNumber2.isSame(number));
     };
 }
