+++ src/main/java/lotto/domain/number/BonusNumber.java
@@ -2,7 +2,7 @@ package lotto.domain.number;
 
 import lotto.domain.lottoForm.WinningNumbers;
 
-public class BonusNumber extends LottoNumber {
+public class BonusNumber extends Number {
     public BonusNumber(int number, WinningNumbers winningNumbers) {
         super(number);
         winningNumbers.validateDuplicate(this);
