+++ src/main/java/lotto/domain/number/LottoNumber.java
@@ -1,29 +1,8 @@
 package lotto.domain.number;
 
-import static lotto.LottoConstants.LOTTO_NUMBER_MAX;
-import static lotto.LottoConstants.LOTTO_NUMBER_MIN;
-import static lotto.message.ErrorMessage.LOTTO_SCOPE_ERROR;
+public class LottoNumber extends Number {
 
-public abstract class LottoNumber {
-    private final int number;
-
-    public LottoNumber(int number) {
-        validateScope(number);
-        this.number = number;
-    }
-
-    private void validateScope(int number) {
-        boolean result = number >= LOTTO_NUMBER_MIN && number <= LOTTO_NUMBER_MAX;
-        if (!result) {
-            throw new IllegalArgumentException(LOTTO_SCOPE_ERROR.getMessage());
-        }
-    }
-
-    public int getNumber() {
-        return number;
-    }
-
-    public boolean isSame(LottoNumber lottoNumber) {
-        return number == lottoNumber.getNumber();
+    public LottoNumber(int number){
+        super(number);
     }
 }
