+++ src/test/java/lotto/domain/LottoTest.java
@@ -2,8 +2,8 @@ package lotto.domain;
 
 import lotto.domain.lottoForm.Lotto;
 import lotto.domain.lottoForm.WinningNumbers;
-import lotto.domain.number.LottoNumber;
 import lotto.domain.number.Number;
+import lotto.domain.number.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -66,12 +66,12 @@ public class LottoTest {
 
         // when
         Lotto lotto = new Lotto(numbers);
-        List<Number> lottoNumbers = lotto.getNumbers();
-        System.out.println(lottoNumbers);
+        List<LottoNumber> lottoLottoNumbers = lotto.getNumbers();
+        System.out.println(lottoLottoNumbers);
 
         // then
         for (int i = 0; i < LOTTO_SIZE - 1; i++) {
-            assertThat(lottoNumbers.get(i).getNumber()).isLessThan(lottoNumbers.get(i + 1).getNumber());
+            assertThat(lottoLottoNumbers.get(i).getNumber()).isLessThan(lottoLottoNumbers.get(i + 1).getNumber());
         }
     }
 
@@ -109,7 +109,7 @@ public class LottoTest {
         List<Integer> numbers = Arrays.stream(lottoInput.split(","))
                 .map(Integer::parseInt)
                 .toList();
-        LottoNumber bonusNumber = new Number(bonusInput);
+        Number bonusNumber = new LottoNumber(bonusInput);
         Lotto lotto = new Lotto(numbers);
 
         // when
